don't bother embedding apfs.efi for setBoot
because we may be using bless on a non-bootable volume that doesn't have apfs.efi - for example, when using gfxutil.sh getefipath command on a non-system volume.

Instead, assume it's been done properly already when macOS was installed to it, or when Startup Disk preferences panel was used to select it as the next boot, or when the real bless command was used to --setBoot.